fix: Add base local conflict processing
